feat: create MobileNav component
