+++ polaris.shopify.com/src/components/MobileNav/index.ts
@@ -1,0 +1,3 @@
+import MobileNav from "./MobileNav";
+
+export default MobileNav;
